Refactors user-db lib into service pattern
